更换了数据库为SQLite
1.重写了数据库操作部分；
2.修复了ssl非正常销毁的漏洞；
3.更新了context.h。
