+++ .gitignore
@@ -6,3 +6,7 @@ connection/
 MySocket-Updater/
 token.h
 json.hpp
+shell.c
+sqlite3.c
+sqlite3.h
+sqlite3ext.h
