+++ README.md
@@ -72,6 +72,10 @@
 6.添加了一处注释(service.cpp)。
 ### 2023.7.30
 进行了代码优化。
+### 2023.8.6
+1.重写了数据库操作部分；  
+2.修复了ssl非正常销毁的漏洞；  
+3.更新了context.h。
 
 ## 待修复bug列表 
 暂无
