--- connection-new.cpp
@@ -1,129 +1,0 @@
-#include "connection-new.h"
-
-using std::map;
-
-static const map<colname,string>GetColname={
-    {colname::id, "ID"},
-    {colname::username, "USERNAME"},
-    {colname::password, "PASSWORD"},
-    {colname::create_at, "CREATE_AT"}
-};
-
-static const map<string, colname>GetColStr={
-    {"ID", colname::id},
-    {"USERNAME", colname::username},
-    {"PASSWORD", colname::password},
-    {"CREATE_AT", colname::create_at}
-};
-
-static int SelectCallback(void* data, int argc, char** argv, char** azColName){
-    map<string, string>* Pack=reinterpret_cast<map<string, string>*>(data);
-    for(int i=0;i<argc;i++){
-        string ColName=string(azColName[i]);
-        string Value=string(argv[i]?argv[i]:"");
-        (*Pack)[ColName]=Value;
-    }
-    return 0;
-}
-
-static int CheckTableCallback(void* data, int argc, char** argv, char** azColName){
-    bool* Flag=reinterpret_cast<bool*>(data);
-    if(argc==0)(*Flag)=false;
-    else (*Flag)=true;
-    return 0;
-}
-
-static int CheckLegalityCallback(void* data, int argc, char** argv, char** azColName){
-    bool* Flag=reinterpret_cast<bool*>(data);
-    int cnt=0;
-    (*Flag)=true;
-    for(int i=0;i<argc;i++){
-        string tmp1=string(azColName[i]);
-        string tmp2=string(argv[i]);
-        if(GetColStr.find(tmp1)!=GetColStr.end())cnt++;
-        if(tmp1=="USERNAME"&&tmp2!="__index")(*Flag)=false;
-        if(tmp1=="PASSWORD"&&tmp2!="__index")(*Flag)=false;
-    }
-    if(cnt<2)(*Flag)=false;
-    else if((*Flag)!=false)(*Flag)=true;
-    return 0;
-}
-
-MyConnection::~MyConnection(){
-    sqlite3_close(db);
-}
-
-string MyConnection::OpenDatabase(const string FileName, const string TableName){
-    bool state=false;
-    int res=sqlite3_open(FileName.c_str(), &db);
-    if(res)return string(sqlite3_errmsg(db));
-    string sql="SELECT name FROM sqlite_master WHERE type='table' AND name='"+TableName+"';";
-    res=sqlite3_exec(db, sql.c_str(), CheckTableCallback, &state, &errMsg);
-    if(res!=SQLITE_OK)return GetErr();
-    if(state)return "";
-    else return "Table is not exist.";
-}
-
-int MyConnection::CreateTable(const string TableName){
-    this->TableName=TableName;
-    string sql=CreateSqlF+TableName+CreateSqlB;
-    int res=sqlite3_exec(db, sql.c_str(), NULL, NULL, &errMsg);
-    if(res!=SQLITE_OK)return res;
-    return Insert("__index", "__index");
-}
-
-int MyConnection::Insert(const string UserName, const string Password){
-    string sql="INSERT INTO TABLE "+TableName+"(USERNAME, PASSWORD)"
-               "VALUE ('"+UserName+"', '"+Password+"');";
-    return sqlite3_exec(db, sql.c_str(), NULL, NULL, &errMsg);
-}
-
-int MyConnection::Select(const colname type, const string str){
-    string sql="SELECT * FROM "+TableName+" WHERE "+GetColname.find(type)->second+" = '"+str+"';";
-    return sqlite3_exec(db, sql.c_str(), SelectCallback, reinterpret_cast<void*>(&Datapack), &errMsg);
-}
-
-int MyConnection::Update(const colname type, const string str){
-    string sql="UPDATE "+TableName+" SET "+GetColname.find(type)->second+" = '"+str+"';";
-    return sqlite3_exec(db ,sql.c_str(), NULL, NULL, &errMsg);
-}
-
-int MyConnection::Delete(const string index){
-    string sql="DELETE FROM "+TableName+" WHERE USERNAME='"+index+"';";
-    return sqlite3_exec(db, sql.c_str(), NULL, NULL, &errMsg);
-}
-
-int MyConnection::Query(const optype type, const string str1, const string str2){
-    if(type==optype::_login){
-        int res=Select(colname::username, str1);
-        if(res!=SQLITE_OK)return res;
-        if(Datapack["PASSWORD"]==str2)return 1;
-        else return 0;
-    }
-    if(type==optype::_register){
-        int res=Select(colname::username, str1);
-        if(res!=SQLITE_OK)return res;
-        if(Datapack["USERNAME"].empty())return 1;
-        else return 0;
-    }
-}
-
-int MyConnection::RebuildTable(){
-    string sql="DROP TABLE "+TableName;
-    int res=sqlite3_exec(db, sql.c_str(), NULL, NULL, &errMsg);
-    if(res!=SQLITE_OK)return res;
-    CreateTable(TableName);
-}
-
-int MyConnection::CheckTable(){
-    string sql="SELECT * FROM "+TableName+" WHERE USERNAME = '__index';";
-    bool state=false;
-    int res=sqlite3_exec(db, sql.c_str(), CheckLegalityCallback, &state, &errMsg);
-    if(res!=SQLITE_OK)return res;
-    if(state==true)return 1;
-    else return 0;
-}
-
-string MyConnection::GetErr(){
-    return string(errMsg);
-}
